bindings/rust/src/lib.rs: move pippenger_mult_impl to a separate file.
